Undo unnecessary import changes, and move added tests away from the test_*_api.py files
